Fix assert, hit for memory operands without base register.
Newer AArch64 instructions can have memory operands without a base register
(only an immediate absolute address).
The instruction '00000098: ldrsw x0, 0' is such a case.
ESIL of course doens't handle these, but also expects always a base register as string.
We can't stringify the disponent in such a case, because the
disp string either leaks or everything has to be
refactored to free it for this case.
This is not worth it, because ESIL is soon gone.
The inaccuracy in analysis will be neglectable I assume.